+++ fhhps/fhhps.py
@@ -6,17 +6,17 @@ Created on Mon Sep 18 09:03:26 2017
 @author: vitorhadad
 """
 
+import warnings
+from warnings import warn
+
+import matplotlib.pyplot as plt
 import numpy as np
-from sklearn.metrics.pairwise import pairwise_kernels
-from sklearn.linear_model import LinearRegression
 import pandas as pd
 from scipy.stats import norm
+from sklearn.linear_model import LinearRegression
+from sklearn.metrics.pairwise import pairwise_kernels
 from sklearn.preprocessing import PolynomialFeatures
-from warnings import warn
 from tqdm import tqdm
-import matplotlib.pyplot as plt
-import warnings
-
 
 
 class NegativeVarianceError(Exception):
